fix(#165): consolidate shared e2e helpers into conftest + condition-based startup
Move clear_canvas, open_context_menu, refresh_vm_card, and
cleanup_non_vm_cards into tests/e2e/conftest.py as the single authoritative
source, and delete the four duplicate/drifted copies scattered across
test_spawn_registry_e2e.py, test_compat.py, test_vm_manager_e2e.py, and
test_vm_manager_real_e2e.py. Each test file now imports the helpers.

Replace the fixed wait_for_timeout(3000) in the page fixture with a
wait_for_function poll on window.__claudeRtsBootComplete, a new observable
signal set at the end of the boot IIFE in static/index.html. Removes
one root-level race that affected every e2e module.

refresh_vm_card now polls /api/vms/favorites and waits until every
favorite has its [data-vm-remove] button in the DOM (empty favorites
accepts [data-vm-search] as the ready signal). Unifies the mock (1.5s)
and real-Docker (3s) fixed sleeps into one condition-based helper.

cleanup_non_vm_cards now uses the condition-based form that waits until
every remaining [data-card-id] contains [data-vm-search].

No test assertions are changed; this is a consolidation + flakiness fix.
