+++ claude_rts/static/index.html
@@ -3925,6 +3925,10 @@ async function handleBlueprintOpenWidget(msg) {
 
   // Center viewport on the cards
   setTimeout(fitAll, 200);
+
+  // Boot complete — observable signal for E2E tests that need a ready
+  // condition instead of a fixed timeout.
+  window.__claudeRtsBootComplete = true;
 })();
 </script>
 </body>
